+++ tests/conftest.py
@@ -79,10 +79,10 @@ def dummy_plugin_config() -> KedroAzureMLConfig:
 @pytest.fixture()
 def patched_kedro_package():
     with patch("kedro.framework.project.PACKAGE_NAME", "tests") as patched_package:
-        original_dir = os.getcwd()
-        os.chdir("tests")
+        # original_dir = os.getcwd()
+        # os.chdir("tests")
         yield patched_package
-        os.chdir(original_dir)
+        # os.chdir(original_dir)
 
 
 @pytest.fixture()