+++ tests/settings.py
@@ -1,1 +1,4 @@
 # Empty file to load `tests` module as Kedro project
+from kedro.config import OmegaConfigLoader  # new import
+
+CONFIG_LOADER_CLASS = OmegaConfigLoader